+++ docs/source/changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.6.11 (2024-XX-XX)
+
+## 0.7.0 (2024-XX-XX)
+- Rework `TableReference` support:
+  * Rename `TableReference` to `ExternalTableReference`
+  * Add support for `ExternalTableReference` to point to tables in external (i.e. not managed by `pipedag`) schemas. 
+  * Remove support for `ExternalTableReference` that points to table in schema of current stage. I.e. `ExternalTableReference` can only point to tables in external schemas.
 - Fix polars import in `pyproject.toml` when using OS X with rosetta2
 
 ## 0.6.10 (2024-02-29)